requirements.txt
fastapi==0.115.4
uvicorn==0.34.3
pydantic==2.12.5
pymupdf4llm==0.2.9
llama-cpp-python==0.3.9
langchain-text-splitters==1.1.0
chromadb==1.5.0
sentence-transformers==4.1.0
huggingface_hub==0.36.2
psutil
requests
python-multipart
